1.20.1 support
Starting with this commit, VF will now support LTS or generally common versions (besides the latest minor of every major version). Many people have requested this in the past and there is no reason to not just provide it, now 1.20.1 users don't have to use ViaProxy if they are using Fabric anyway.